+++ ChangeLog.md
@@ -6,7 +6,7 @@
 + Add infors (CODE_OF_CONDUCT.md, MANIFEST.in, LICENSE, requirements.txt)
 + Add shared modules (activator, boundary_controller, data_processor, randomer, scaler, and scorer)
 + Add `RvflRegressor` and `RvflClassifier` classes
-+ Add `MhaTuneRvfl` class that can be used to tune RVFL network
++ Add `GfoRvflTuner` class that can be used to tune RVFL network
 + Add publish workflow
 + Add examples and tests folders